requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
astroid==2.15.6
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
dill==0.3.7
frozenlist==1.3.3
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
multidict==6.0.4
openai==0.27.8
platformdirs==3.10.0
prompt-toolkit==3.0.38
pylint==2.17.5
python-dotenv==1.0.0
requests==2.31.0
tomlkit==0.12.1
tqdm==4.65.0
urllib3==2.0.3
wcwidth==0.2.6
wrapt==1.15.0
yarl==1.9.2
